feat: create list villages in sidebar maps
